refactor: reorganize animation scripts into global and per-page directories
