+++ src/components/global/NavbarAlt.astro
@@ -123,7 +123,7 @@ const currentPath = Astro.url.pathname;
 </header>
 
 <script>
-  import { nabBarAltAnimation } from "@/utils/scripts/nabBarAltAnimation.js";
+  import { nabBarAltAnimation } from "@/utils/scripts/globalAnimations/nabBarAltAnimation.js";
 
   document.removeEventListener("DOMContentLoaded", nabBarAltAnimation);
   document.addEventListener("DOMContentLoaded", nabBarAltAnimation);
