+++ src/components/section/landing/LatestProjectsAlt.astro
@@ -154,7 +154,7 @@ const sortedWorks = allworks.sort(
 </section>
 
 <script>
-  import { latestProjectsAltAnimation } from "@/utils/scripts/latestProjectsAltAnimation.js";
+  import { latestProjectsAltAnimation } from "@/utils/scripts/perPageAnimations/home/latestProjectsAltAnimation.js";
 
   document.removeEventListener("DOMContentLoaded", latestProjectsAltAnimation);
   document.addEventListener("DOMContentLoaded", latestProjectsAltAnimation);
