+++ src/components/section/landing/ServicesLanding.astro
@@ -63,7 +63,7 @@ import Button from "@/components/ui/button.astro";
 </section>
 
 <script>
-  import { servicesAltAnimation } from "@/utils/scripts/servicesAltAnimation.js";
+  import { servicesAltAnimation } from "@/utils/scripts/perPageAnimations/home/servicesAltAnimation.js";
 
   document.removeEventListener("DOMContentLoaded", servicesAltAnimation);
   document.addEventListener("DOMContentLoaded", servicesAltAnimation);
