+++ src/components/section/landing/Solution.astro
@@ -51,7 +51,7 @@ const { className } = Astro.props;
 </section>
 
 <script>
-  import { solutionsAnimation } from "@/utils/scripts/solutionsAnimation.js";
+  import { solutionsAnimation } from "@/utils/scripts/perPageAnimations/home/solutionsAnimation.js";
   document.removeEventListener("DOMContentLoaded", solutionsAnimation);
   document.addEventListener("DOMContentLoaded", solutionsAnimation);
 </script>
