+++ src/components/section/services/ServicesStepsProcess.astro
@@ -11,10 +11,12 @@ const { titleLine1, titleLine2, service } = Astro.props;
 ---
 
 <section id="process" class="max-w-6xl mx-auto py-10 lg:py-20 px-4">
-  <h2 class="title text-center text-[min(10vw,42px)] font-playfair opacity-0 mb-0">
-    <span class="block">{titleLine1}</span>
-    <span class="">{titleLine2}</span>
-  </h2>
+  <div class="process-heading opacity-0">
+    <h2 class="text-center text-[min(10vw,42px)] font-playfair mb-0">
+      <span class="block">{titleLine1}</span>
+      <span class="">{titleLine2}</span>
+    </h2>
+  </div>
 
   <!-- SVG Progress Bar - Hidden on mobile -->
   <div class="hidden lg:block -mt-8">
@@ -135,7 +137,7 @@ const { titleLine1, titleLine2, service } = Astro.props;
 </section>
 
 <script>
-  import { gsapProcessStepsAnimation } from "@/utils/scripts/servicesStepsProccesAnimation";
+  import { gsapProcessStepsAnimation } from "@/utils/scripts/perPageAnimations/servicesAnimations/servicesStepsProccesAnimation";
   document.addEventListener("DOMContentLoaded", () => {
     gsapProcessStepsAnimation();
   });
