+++ src/pages/servicios/marca-personal.astro
@@ -3,7 +3,6 @@ import ServiceLayout from "@/layouts/ServiceLayout.astro";
 import { getServiceBySlug, type Service } from "@/config/services";
 import CTABanner from "@/components/ui/CTABanner.astro";
 import GlobalHero from "@/components/global/GlobalHero.astro";
-import SocialContentStruggled from "@/assets/images/socialContentStruggled.webp";
 import Button from "@/components/ui/button.astro";
 import StarIcon from "@/assets/icons/star40x40.svg";
 import CheckIcon from "@/assets/icons/check.svg";
