+++ src/utils/scripts/globalAnimations/bgAnimation.js
