+++ src/utils/scripts/globalAnimations/nabBarAltAnimation.js
