+++ src/utils/scripts/perPageAnimations/about/aboutHeroAnimation.js
